Final FileManager changes, Buddy List support: LWC-1551

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/CoreRandomGlue.java
@@ -176,7 +176,8 @@ class CoreRandomGlue {
         File tFile;
         for(int i = 0; i < file_list.length; i++) {
             tFile = file_list[i];
-            if(!fileManager.getSharedFileList().contains(tFile) &&
+            if(fileManager.getFileDesc(tFile) != null &&
+                    !fileManager.getSharedFileList().contains(fileManager.getFileDesc(tFile)) &&
                     tFile.lastModified() < purgeLimit) {
                 tFile.delete();
             }
